Switch a few places to using Tessellation instead of constructing an invalid Polygon only to render it. #951
