+++ abstutil/src/serde.rs
@@ -1,6 +1,5 @@
 use std::cmp::Ord;
 use std::collections::{BTreeMap, HashMap};
-use std::convert::TryFrom;
 
 use anyhow::Result;
 use serde::de::DeserializeOwned;
