+++ map_model/src/objects/intersection.rs
@@ -1,5 +1,4 @@
 use std::collections::{BTreeMap, BTreeSet};
-use std::convert::TryFrom;
 use std::fmt;
 
 use serde::{Deserialize, Serialize};
